Breaking apart converters
